Fix path new

+++ admin/send_email.py
@@ -9,24 +9,18 @@ from email.mime.text import MIMEText
 from cmdline_add_args.allure_report_path import CURRENT_DIRECTORY
 from dotenv import load_dotenv
 
-from admin.allure_env import get_os_type
+from admin.logger import logger
 from config.env import BROWSER, SEND_REPORT
 
 load_dotenv()
 
-report_path = CURRENT_DIRECTORY + f'/allure-results'
-
-if get_os_type() == "Windows":
-    path = report_path.replace("/", "\\")
-else:
-    path = report_path
-
-target_path = path
-file_to_zip = BROWSER
-ADMIN_FOLDER = os.path.dirname(__file__)
+report_path = os.path.join(CURRENT_DIRECTORY, 'allure-results')
+target_path = os.path.normpath(report_path)
+file_to_zip = 'IPAD 10.6.0'
 
 output_filename = f'{BROWSER} report'
-output_path = ADMIN_FOLDER
+output_path = os.path.dirname(__file__)
+filename = f"{output_filename}.zip"
 
 # Mail settings
 from_email = os.getenv('FROM_EMAIL')
@@ -34,32 +28,32 @@ password = os.getenv('PASSWORD')
 to_email = os.getenv('TO_EMAIL')
 subject = "Allure report"
 body = f"{BROWSER} report"
-filename = f"{output_filename}.zip"
 
 
 class SendMail:
     def __init__(self, target_path, file_to_zip):
-        # Archive report
+        self.target_path = target_path
+        self.file_to_zip = file_to_zip
+        self.archive_report()
+
+    def archive_report(self):
         try:
-            shutil.make_archive(os.path.join(ADMIN_FOLDER, output_filename), 'zip', target_path, file_to_zip)
-        except:
-            print('pass')
+            shutil.make_archive(os.path.join(output_path, output_filename), 'zip', self.target_path, self.file_to_zip)
+        except Exception as e:
+            print(f"Error creating archive: {e}")
 
     # Attaching a file
     def attaching_a_file(self):
-        attachment_file = f"{output_path}\\{filename}"
-        if get_os_type() == "Windows":
-            correct_path = attachment_file
-        else:
-            correct_path = attachment_file.replace("\\", "/")
+        attachment_file = os.path.join(output_path, filename)
+        correct_path = os.path.normpath(attachment_file)
         try:
             with open(correct_path, 'rb') as attachment:
                 part = MIMEBase('application', 'octet-stream')
                 part.set_payload(attachment.read())
                 encoders.encode_base64(part)
-                part.add_header('Content-Disposition', f"attachment; filename= {filename}")
+                part.add_header('Content-Disposition', f"attachment; filename={filename}")
                 return part
-        except UnicodeDecodeError as e:
+        except Exception as e:
             print(f"Error reading file {correct_path}: {e}")
             return None
 
@@ -70,7 +64,7 @@ class SendMail:
         msg['To'] = to_email
         msg['Subject'] = subject
         msg.attach(MIMEText(body, 'plain'))
-        attachment_part = SendMail.attaching_a_file(self)
+        attachment_part = self.attaching_a_file()
         if attachment_part:
             msg.attach(attachment_part)
         try:
@@ -78,18 +72,24 @@ class SendMail:
                 server.starttls()
                 server.login(from_email, password)
                 server.sendmail(from_email, to_email, msg.as_string())
-
             print("Email sent successfully")
         except smtplib.SMTPAuthenticationError:
             print("\033[91mAuthentication unsuccessful, the user credentials were incorrect\033[0m")
+        except Exception as e:
+            print(f"Error sending email: {e}")
         finally:
             archive_file = os.path.join(output_path, filename)
-            os.remove(archive_file)
+            if os.path.exists(archive_file):
+                os.remove(archive_file)
 
 
 def send_report_to_email():
+    logger.info(f"SEND REPORT: {SEND_REPORT}")
     if SEND_REPORT:
         print("\nSend report")
         print("Slack zip: ", filename)
-        SendMail(target_path, file_to_zip).send_mail()
+        SendMail(target_path, filename).send_mail()
 
+
+if __name__ == '__main__':
+    send_report_to_email()
